web_view_safetensor_model/requirements.txt
Flask==2.3.3
safetensors==0.4.1
torch==2.1.0
networkx==3.2.1
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
